1.4.0 product filter (#1692)
* novator: ACORN-969 novator filtration. Category page.

* novator: ACORN-969 novator filtration. Brand page.

* admin: ant request fix

* novator: ACORN-969 novator filtration. Search page.

* novator: ACORN-969 novator filtration. Search page.

+++ public_html/admin/controller/responses/common/do_embed.php
@@ -171,7 +171,7 @@ class ControllerResponsesCommonDoEmbed extends AController
             $subcategories = [];
             //if embed for only one category
             if (sizeof($category_id) == 1) {
-                $cat_id = current($category_id);
+                $cat_id = (int)current($category_id);
                 $category_info = $this->model_catalog_category->getCategory($cat_id);
                 $subcategories = $this->model_catalog_category->getCategories($cat_id);
                 if ($category_info['parent_id'] == 0) {
